--- examples/hot-reload-tide/tests/messages_test_config.json
@@ -1,9 +1,0 @@
-{
-  "audio_folder_path": "sounds/",
-  "messages": {
-    "sound.mp3": {
-      "display_name": "Sound 1",
-      "volume": 60.0
-    }
-  }
-}
