+++ examples/hot_reload_tide/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "announcer"
+name = "hot-reload-tide"
 version = "0.1.0"
 authors = ["user <savovs@gmail.com>"]
 edition = "2018"
